Refactor builder class
