+++ inngest/src/main/kotlin/com/inngest/InngestFunction.kt
@@ -38,64 +38,8 @@ abstract class InngestFunction {
 
 //    abstract val id: String;
 
-    // TODO: Throw illegal argument exception
-    class Builder(val id: String) {
-        private var name: String? = null;
-        private var triggers: MutableList<InngestFunctionTrigger> = mutableListOf();
-        private var batchEvents: BatchEvents? = null;
 
-        fun name(name: String): Builder {
-            this.name = name
-            return this
-        }
-
-        fun trigger(trigger: InngestFunctionTrigger): Builder {
-            // TODO - Check max triggers
-            // TODO - Check mutually exclusive opts (cron v. event+if?)
-            this.triggers.add(trigger)
-            return this
-        }
-
-        fun batchEvents(maxSize: Int, timeout: Duration, key: String? = null): Builder {
-            this.batchEvents = BatchEvents(maxSize, timeout, key)
-            return this;
-        }
-
-        private fun buildSteps(serveUrl: String): Map<String, StepConfig> {
-            val scheme = serveUrl.split("://")[0]
-            return mapOf(
-                "step" to
-                    StepConfig(
-                        id = "step",
-                        name = "step",
-                        retries =
-                        mapOf(
-                            // TODO - Pull from conf option
-                            "attempts" to 3,
-                        ),
-                        runtime =
-                        hashMapOf(
-                            "type" to scheme,
-                            "url" to "$serveUrl?fnId=${id}&stepId=step",
-                        ),
-                    ),
-            )
-        }
-
-        fun build(appId: String, serverUrl: String): InternalFunctionConfig {
-            val globalId = String.format("%s-%s", appId, id)
-            val config = InternalFunctionConfig(
-                globalId,
-                name,
-                triggers,
-                batchEvents,
-                steps = buildSteps(serverUrl)
-            )
-            return config
-        }
-    }
-
-    open fun config(builder: Builder): Builder {
+    open fun config(builder: InngestFunctionConfigBuilder): InngestFunctionConfigBuilder {
         return builder
     }
 
@@ -118,7 +62,7 @@ abstract class InngestFunction {
 //            throw Exception("FunctionConfig annotation is required to setup an InngestFunction")
 //        }
         val triggers = buildEventTriggers() + buildCronTriggers() + buildIfTriggers()
-        val builder = Builder(id())
+        val builder = InngestFunctionConfigBuilder()
         val configBuilder = this.config(builder)
         return InternalInngestFunction(configBuilder, this::execute)
     }
@@ -137,32 +81,5 @@ abstract class InngestFunction {
             .map { InternalFunctionTrigger(event = (it as FunctionIfTrigger).`if`) }
 }
 
-data class InngestFunctionTrigger
-@JvmOverloads
-constructor(
-    @Json(serializeNull = false) val event: String? = null,
-    @Json(name = "expression", serializeNull = false) val `if`: String? = null,
-    @Json(serializeNull = false) val cron: String? = null,
-)
 
-@Target(AnnotationTarget.FIELD)
-annotation class KlaxonDuration
 
-val durationConverter = object : Converter {
-    override fun canConvert(cls: Class<*>): Boolean = cls == Duration::class.java;
-
-    // TODO Implement this - parse 30s into duration of seconds
-    override fun fromJson(jv: JsonValue): Duration =
-        throw KlaxonException("Duration parse not implemented: ${jv.string}")
-
-    override fun toJson(value: Any): String = """"${(value as Duration).seconds}s""""
-}
-
-data class BatchEvents
-@JvmOverloads
-constructor(
-    val maxSize: Int,
-    @KlaxonDuration
-    val timeout: Duration,
-    @Json(serializeNull = false) val key: String? = null
-)
